refactor: Updated settings and render functions
